requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
distlib==0.3.1
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
mysql-connector-python==8.0.24
numpy==1.20.2
pandas==1.2.4
protobuf==3.15.8
psycopg2==2.8.6

pylint==2.7.4
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
six==1.15.0
snakeviz==2.1.0
soupsieve==2.2.1
SQLAlchemy==1.4.11
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
urllib3==1.26.4
virtualenv==20.4.3
whitenoise==5.2.0
wrapt==1.12.1
